1. 修复createAbsMv中conMvs收集错误,导致absMv未存储的bug;   2. absDelta计算有误的bugfix;   3. 补充mv可视化的"点击描述";

+++ SMG_NothingIsAll/AIFoundation/AINet/AIAbsCMVManager/AINetAbsCMVUtil.m
@@ -64,7 +64,7 @@
     }
     
     //3. 取absUrgentTo | absDelta; (//由MIN(aUrgentTo, bUrgentTo)改为平均)
-    NSInteger absValue = sum / mvNodes.count;
+    NSInteger absValue = sum / (int)mvNodes.count;
     return absValue;
 }
